+++ esphome/components/ade7953/sensor.py
@@ -1,90 +1,14 @@
-import esphome.codegen as cg
-import esphome.config_validation as cv
-from esphome.components import sensor, i2c
-from esphome import pins
-from esphome.const import (
-    CONF_ID,
-    CONF_IRQ_PIN,
-    CONF_VOLTAGE,
-    DEVICE_CLASS_CURRENT,
-    DEVICE_CLASS_POWER,
-    DEVICE_CLASS_VOLTAGE,
-    STATE_CLASS_MEASUREMENT,
-    UNIT_AMPERE,
-    UNIT_VOLT,
-    UNIT_WATT,
-)
+import sys
+from esphome.components.ade7953_i2c import sensor as ade7953_i2c
+sys.path.insert(0,"..") 
 
-DEPENDENCIES = ["i2c"]
+# ade7953 is an alias of ade7953_i2c. ade7953 was used before
+# support for spi was added
 
-ade7953_ns = cg.esphome_ns.namespace("ade7953")
-ADE7953 = ade7953_ns.class_("ADE7953", cg.PollingComponent, i2c.I2CDevice)
+DEPENDENCIES = ade7953_i2c.DEPENDENCIES
+AUTO_LOAD = ade7953_i2c.AUTO_LOAD + ["ade7953_i2c"]
 
-CONF_CURRENT_A = "current_a"
-CONF_CURRENT_B = "current_b"
-CONF_ACTIVE_POWER_A = "active_power_a"
-CONF_ACTIVE_POWER_B = "active_power_b"
-
-CONFIG_SCHEMA = (
-    cv.Schema(
-        {
-            cv.GenerateID(): cv.declare_id(ADE7953),
-            cv.Optional(CONF_IRQ_PIN): pins.internal_gpio_input_pin_schema,
-            cv.Optional(CONF_VOLTAGE): sensor.sensor_schema(
-                unit_of_measurement=UNIT_VOLT,
-                accuracy_decimals=1,
-                device_class=DEVICE_CLASS_VOLTAGE,
-                state_class=STATE_CLASS_MEASUREMENT,
-            ),
-            cv.Optional(CONF_CURRENT_A): sensor.sensor_schema(
-                unit_of_measurement=UNIT_AMPERE,
-                accuracy_decimals=2,
-                device_class=DEVICE_CLASS_CURRENT,
-                state_class=STATE_CLASS_MEASUREMENT,
-            ),
-            cv.Optional(CONF_CURRENT_B): sensor.sensor_schema(
-                unit_of_measurement=UNIT_AMPERE,
-                accuracy_decimals=2,
-                device_class=DEVICE_CLASS_CURRENT,
-                state_class=STATE_CLASS_MEASUREMENT,
-            ),
-            cv.Optional(CONF_ACTIVE_POWER_A): sensor.sensor_schema(
-                unit_of_measurement=UNIT_WATT,
-                accuracy_decimals=1,
-                device_class=DEVICE_CLASS_POWER,
-                state_class=STATE_CLASS_MEASUREMENT,
-            ),
-            cv.Optional(CONF_ACTIVE_POWER_B): sensor.sensor_schema(
-                unit_of_measurement=UNIT_WATT,
-                accuracy_decimals=1,
-                device_class=DEVICE_CLASS_POWER,
-                state_class=STATE_CLASS_MEASUREMENT,
-            ),
-        }
-    )
-    .extend(cv.polling_component_schema("60s"))
-    .extend(i2c.i2c_device_schema(0x38))
-)
-
+CONFIG_SCHEMA = ade7953_i2c.CONFIG_SCHEMA
 
 async def to_code(config):
-    var = cg.new_Pvariable(config[CONF_ID])
-    await cg.register_component(var, config)
-    await i2c.register_i2c_device(var, config)
-
-    if irq_pin_config := config.get(CONF_IRQ_PIN):
-        irq_pin = await cg.gpio_pin_expression(irq_pin_config)
-        cg.add(var.set_irq_pin(irq_pin))
-
-    for key in [
-        CONF_VOLTAGE,
-        CONF_CURRENT_A,
-        CONF_CURRENT_B,
-        CONF_ACTIVE_POWER_A,
-        CONF_ACTIVE_POWER_B,
-    ]:
-        if key not in config:
-            continue
-        conf = config[key]
-        sens = await sensor.new_sensor(conf)
-        cg.add(getattr(var, f"set_{key}_sensor")(sens))
+    await ade7953_i2c.to_code(config)
